refactor esri file upload, move geometry preview to separate controller
